requirements.txt
Django==3.2.12
Django: Web framework for building the application.

pandas==1.3.5
pandas: Data manipulation library to handle Excel files.

djangorestframework==3.13.1
djangorestframework: Django REST framework for building APIs.

openpyxl==3.0.9
openpyxl: Library to read and write Excel files with Python.